requirements.txt
streamlit
google-generativeai
protobuf
requests
uritemplate
